Update Portfile to latest release
I am not sure how to fix the checksums.

This install should set caddy environmental variables to be similar to other macports installs.  i.e.:
- caddy.HomeDir= /opt/local/var/www/    (or whatever the apache2 web base is)
- caddy.AppDataDir = /opt/local/var/run/caddy/
- caddy.AppConfigDir = /opt/local/etc/caddy/
- caddy.ConfigAutosavePath=/opt/local/var/run/caddy/autosave.json

I have been poking around the macports guides but have not been able to finish how to build a portfile.  The biggest reasons to normalize the environmental folders is that caddy can manage cert files.  by default, it will want to put these in the ~/library/application support/...     normalizing it into the macports schema (/opt/local/) allows the use of the certs it manages by other ports and does it in a sane way where the security can be managed.

Also, caddy has trouble binding privileged ports unless run with privilege.   Theoretically, the macOS way is to use the helper application, but starting caddy with root gives it what it needs, so long as there is a chmod +s on the binary.   I have a launchctl plist that runs caddy with privilege and then sets it to run as user www:www

This allows caddy to work on the same files as apache/nginx/php-fpm. Makes it a good citizen in the web server groups, and normalizes file permissions.

I can post the launchctl plist as a gist if someone can help with the portfile.

I see other macports have a 'port load xxx' that loads the service.   If I knew how to add this to this port, I would.